+++ src/other_plots.jl
@@ -1,0 +1,84 @@
+#
+
+## Rateplot
+
+function rateplot(p, sym)
+    r = getrecord(p, sym)
+    R = hcat(r...)
+end
+
+function rateplot(P::Array, sym)
+    R = vcat([rateplot(p, sym) for p in P]...)
+    y0 = [p.N for p in P][2:(end-1)]
+    plt = heatmap(R, leg = :none)
+    !isempty(y0) && hline!(plt, cumsum(y0), line = (:black, 1))
+    plt
+end
+
+
+function if_curve(model, current; neuron = 1, dt = 0.1ms, duration = 1second)
+    E = model(neuron)
+    monitor!(E, [:fire])
+    f = Float32[]
+    for I in current
+        clear_records!(E)
+        E.I = [I]
+        sim!([E], []; dt = dt, duration = duration)
+        push!(f, activity(E))
+    end
+    plot(current, f)
+end
+
+export rateplot
+
+# export density
+# function density(p, sym)
+#   X = getrecord(p, sym)
+#   t = dt*(1:length(X))
+#   xmin, xmax = extrema(vcat(X...))
+#   edge = linspace(xmin, xmax, 100)
+#   c = center(edge)
+#   ρ = [fit(Histogram, x, edge).weights |> reverse |> float for x in X] |> x->hcat(x...)
+#   ρ = smooth(ρ, windowsize(p), 2)
+#   ρ ./= sum(ρ, 1)
+#   surface(t, c, ρ, ylabel="p")
+# end
+# function density(P::Array, sym)
+#   plts = [density(p, sym) for p in P]
+#   plot(plts..., layout=(length(plts),1))
+# end
+
+
+# function windowsize(p)
+#     A = sum.(p.records[:fire]) / length(p.N)
+#     W = round(Int32, 0.5p.N / mean(A)) # filter window, unit=1
+# end
+
+# function density(p, sym)
+#     X = getrecord(p, sym)
+#     t = 1:length(X)
+#     xmin, xmax = extrema(vcat(X...))
+#     edge = linspace(xmin, xmax, 50)
+#     c = center(edge)
+#     ρ = [fit(Histogram, x, edge).weights |> float for x in X] |> x -> hcat(x...)
+#     ρ = smooth(ρ, windowsize(p), 2)
+#     ρ ./= sum(ρ, 1)
+#     p = @gif for t = 1:length(X)
+#         bar(c, ρ[:, t], leg = false, xlabel = string(sym), yaxis = ("p", extrema(ρ)))
+#     end
+#     is_windows() && run(`powershell start $(p.filename)`)
+#     is_unix() && run(`xdg-open $(p.filename)`)
+#     p
+# end
+
+# function activity(p)
+#     A = sum.(p.records[:fire]) / length(p.N)
+#     W = windowsize(p)
+#     A = smooth(A, W)
+# end
+
+# function activity(P::Array)
+#     A = activity.(P)
+#     t = 1:length(P[1].records[:fire])
+#     plot(t, A, leg = :none, xaxis = ("t",), yaxis = ("A", (0, Inf)))
+# end
